frontend_service/requirements.txt
fastapi
uvicorn[standart]
kafka-python
pydantic-settings
